+++ GUI/ResponseTimeInfo.java
@@ -7,7 +7,7 @@ public class ResponseTimeInfo extends JLabel {
 
     private final Color white = new java.awt.Color(166, 166, 166);
 
-    public ResponseTimeInfo(int time, Color colorOfForground){
+    public ResponseTimeInfo(Long time, Color colorOfForground){
         super(time+"ms", SwingConstants.CENTER);
         this.setPreferredSize(new Dimension(40, 30));
         this.setBorder(BorderFactory.createLineBorder(new java.awt.Color(218, 218, 218)));
